refactor: 归位角色领域实现
